Provides Initial Functionality for Lining Up Axis Ticks and Grid Lines
This function exposes the protested member function "next_tick" in
Cartesian to allow for the vector of tick line locations can be
identified. This information can then be set to Viewport, where vectors
are added to keep track of the grid lines in X and Y. Finally,
drawGridOverlay is updated to use the X and Y grid line values if they
are defined, but otherwise use the existing default spacing.